Move bootstrapDefaultConfigSet to ConfigSetService
